refactor util package to hash package
