+++ internal/hash/hash.go
@@ -1,4 +1,5 @@
-package util
+// Package hash handles utility functions for hashing.
+package hash
 
 import (
 	"crypto/sha256"
@@ -6,6 +7,7 @@ import (
 	"strings"
 )
 
+// GenerateSHA256Hex generates a SHA256 from source.
 func GenerateSHA256Hex(source string) string {
 	hash := sha256.New()
 	hash.Write([]byte(source))
